Change sendOffer to sendSdp, and safeParse to parse.
* The websocket is agnostic of whether the SDP is an offer or an answer.
  So the function sending this over the websocket was moved a level
  higher, so that it can be used in the next steps.

* The ws event payload handler functions were migrated from using
  safeParse to parse. Since the occurence of an error is not a part of
  the normal flow of the program, it's best if these errors are thrown
  by zod library, instead of manually in the code.
