+++ src/scripts/pages/room/startSession/connectRoomWs/events/onMessage/createConnectionAndSendOffer.ts
@@ -1,7 +1,7 @@
 import connectionsManager from "scripts/pages/room/startSession/connections/manager";
 import RoomValidators, { RoomId } from "utils/types/Room";
 import { z } from "zod";
-import sendOfferToRoomWs from "./sendOffer";
+import sendSdpToRoomWs from "./_utils/sendSdp";
 
 const PayloadValidator = z.object({
   newMemberId: RoomValidators.RoomMemberId,
@@ -16,14 +16,15 @@ async function createConnectionAndSendOffer({
   roomWs: WebSocket;
   roomId: RoomId;
 }) {
-  const validation = PayloadValidator.safeParse(payload);
-  if (!validation.success)
-    throw new Error("'sendSdp' payload did not contain the expected data");
-  const { newMemberId } = validation.data;
-
+  const { newMemberId } = PayloadValidator.parse(payload);
   const pc = connectionsManager.createConnection(newMemberId);
   const offerSdp = await pc.createOffer();
-  sendOfferToRoomWs({ roomWs, offerSdp, newMemberId, roomId });
+  sendSdpToRoomWs({
+    roomWs,
+    sdp: offerSdp,
+    receiverMemberId: newMemberId,
+    roomId,
+  });
 }
 
 export default createConnectionAndSendOffer;
